Refactored code & Introduced CSV Builder & Factory to ensure programming for Interface then Implementation
