+++ src/main/java/csvBuilder/CSVBuilderFactory.java
@@ -1,0 +1,7 @@
+package csvBuilder;
+
+public class CSVBuilderFactory {
+	public static ICSVBuilder createCSVBuilder() {
+		return (ICSVBuilder) new OpenCSVBuilder();
+	}
+}
